[move]: Moved files to new directory structure for better organization.
