+++ lib/src/cores/router/router.dart
@@ -25,7 +25,7 @@ import 'package:riverpod_annotation/riverpod_annotation.dart';
 
 import '../../utilities/components/greenfield_images_detail.dart';
 import '../../viewmodels/notice/notice_view_model.dart';
-import '../../viewmodels/recruit_view_model.dart';
+import '../../viewmodels/recruit/recruit_view_model.dart';
 import '../../views/setting/setting_view.dart';
 
 part 'router.g.dart';
@@ -202,71 +202,71 @@ GoRouter goRouter(Ref ref) {
               ),
             ],
           ),
-          StatefulShellBranch(
-            navigatorKey: _recruitTabNavigatorKey,
-            routes: [
-              GoRoute(
-                path: '/recruit',
-                builder: (BuildContext context, GoRouterState state) {
-                  return RecruitView();
-                },
-                routes: <RouteBase>[
-                  GoRoute(
-                    name: 'recruit_detial',
-                    path: 'detail/:id',
-                    pageBuilder: (context, state) {
-                      return CustomTransitionPage(
-                        key: state.pageKey,
-                        child: RecruitDetailView(
-                            recruit: recruitVM
-                                .getRecruitById(state.pathParameters['id']!)),
-                        transitionsBuilder:
-                            (context, animation, secondaryAnimation, child) {
-                          const begin = Offset(0.0, 1.0); // 아래에서 시작
-                          const end = Offset.zero; // 현재 위치
-                          const curve = Curves.easeInOut;
-
-                          var tween = Tween(begin: begin, end: end)
-                              .chain(CurveTween(curve: curve));
-                          var offsetAnimation = animation.drive(tween);
-
-                          return SlideTransition(
-                            position: offsetAnimation,
-                            child: child,
-                          );
-                        },
-                      );
-                    },
-                  ),
-                  GoRoute(
-                    name: 'recruit_edit',
-                    path: 'edit',
-                    pageBuilder: (context, state) {
-                      return CustomTransitionPage(
-                        key: state.pageKey,
-                        child: RecruitEditView(),
-                        transitionsBuilder:
-                            (context, animation, secondaryAnimation, child) {
-                          const begin = Offset(0.0, 1.0); // 아래에서 시작
-                          const end = Offset.zero; // 현재 위치
-                          const curve = Curves.easeInOut;
-
-                          var tween = Tween(begin: begin, end: end)
-                              .chain(CurveTween(curve: curve));
-                          var offsetAnimation = animation.drive(tween);
-
-                          return SlideTransition(
-                            position: offsetAnimation,
-                            child: child,
-                          );
-                        },
-                      );
-                    },
-                  ),
-                ],
-              )
-            ],
-          ),
+          // StatefulShellBranch(
+          //   navigatorKey: _recruitTabNavigatorKey,
+          //   routes: [
+          //     GoRoute(
+          //       path: '/recruit',
+          //       builder: (BuildContext context, GoRouterState state) {
+          //         return RecruitView();
+          //       },
+          //       routes: <RouteBase>[
+          //         GoRoute(
+          //           name: 'recruit_detial',
+          //           path: 'detail/:id',
+          //           pageBuilder: (context, state) {
+          //             return CustomTransitionPage(
+          //               key: state.pageKey,
+          //               child: RecruitDetailView(
+          //                   recruit: recruitVM
+          //                       .getRecruitById(state.pathParameters['id']!)),
+          //               transitionsBuilder:
+          //                   (context, animation, secondaryAnimation, child) {
+          //                 const begin = Offset(0.0, 1.0); // 아래에서 시작
+          //                 const end = Offset.zero; // 현재 위치
+          //                 const curve = Curves.easeInOut;
+          //
+          //                 var tween = Tween(begin: begin, end: end)
+          //                     .chain(CurveTween(curve: curve));
+          //                 var offsetAnimation = animation.drive(tween);
+          //
+          //                 return SlideTransition(
+          //                   position: offsetAnimation,
+          //                   child: child,
+          //                 );
+          //               },
+          //             );
+          //           },
+          //         ),
+          //         GoRoute(
+          //           name: 'recruit_edit',
+          //           path: 'edit',
+          //           pageBuilder: (context, state) {
+          //             return CustomTransitionPage(
+          //               key: state.pageKey,
+          //               child: RecruitEditView(),
+          //               transitionsBuilder:
+          //                   (context, animation, secondaryAnimation, child) {
+          //                 const begin = Offset(0.0, 1.0); // 아래에서 시작
+          //                 const end = Offset.zero; // 현재 위치
+          //                 const curve = Curves.easeInOut;
+          //
+          //                 var tween = Tween(begin: begin, end: end)
+          //                     .chain(CurveTween(curve: curve));
+          //                 var offsetAnimation = animation.drive(tween);
+          //
+          //                 return SlideTransition(
+          //                   position: offsetAnimation,
+          //                   child: child,
+          //                 );
+          //               },
+          //             );
+          //           },
+          //         ),
+          //       ],
+          //     )
+          //   ],
+          // ),
           StatefulShellBranch(
             navigatorKey: _postTabNavigatorKey,
             routes: [
