+++ lib/src/datas/repositories/home_repository.dart
@@ -4,7 +4,7 @@ import 'package:green_field/src/datas/services/firebase_auth_service.dart';
 import 'package:firebase_auth/firebase_auth.dart' as firebase_auth;
 import '../../cores/error_handler/result.dart';
 import '../../model/user.dart' as myUser;
-import '../services/firebase_store_service.dart';
+import '../services/firebase_stores/firebase_store_service.dart';
 
 class HomeRepository {
   final FirebaseAuthService firebaseAuthService;
